Don't animate left & top properties
I accidentally inlined these in a previous commit, thinking if they were in the same d3 chain of methods, it would be a no-op.

But the .transition().duration(200) makes these properties animate, leading to a funny dance where the tooltip chases the pointer, 200ms behind.

Moving these property setting before the transition() fixes the problem.